refactor: use a specific function for each right

+++ packages/agent/src/agent/routes/modification/delete.ts
@@ -17,11 +17,7 @@ export default class DeleteRoute extends CollectionRoute {
   }
 
   public async handleDelete(context: Context): Promise<void> {
-    await this.services.authorization.assertCanOnCollection(
-      context,
-      CollectionActionEvent.Delete,
-      this.collection.name,
-    );
+    await this.services.authorization.assertCanDelete(context, this.collection.name);
 
     const id = IdUtils.unpackId(this.collection.schema, context.params.id);
     await this.deleteRecords(context, { ids: [id], areExcluded: false });
